Add enableP2P flag to Node.run and NonP2P functions
The enableP2P flag is to make the correct DiffusionApplications data
type. However in order to do that I had to add a bunch of NonP2P master
branch functions.

I tried to be thorough and addd TODOs to mark that the added functions,
in order to be easier to revert back to p2p-master.

Fixed compiling warnings

Updated consensus Node.hs with latest p2p-master

Refactored Boolean to NetworkP2PMode